🐛 fix linked chart previews not working on the /latest page
